sys: use HKEYI hook instead of HTIMI one
HTIMI is called after clearing the VDP INT flag,
which means we can have reentrant interrupts if
we too much work in interrupt context. If that
happens, the result would be a stack overflow;
but with this change, too much work during
interrupt translates in a stall of the main
cpu thread.