map2header: fix issues related to data inside unions
data from differnt objects is stored as a union struct,
which means we need to add padding in some of the types,
but also that the data needs to be properly ordered in
the output even if the input is random.

the result of this is a rather ineficient packing of the
datam evem though is very easy to access from code.