tileblit: initial implementation and test
The tileblit module allows performing blit operations using multicolor
tiles to superimpose tiles with transparent backgrounds on top of other
tiles. The algorith is currently a bit rough, but it gives an idea of
how this should work: pick the best combination of foreground and
background colors to fill-in the transparent portions of the top tile.

Although this is not necessarily intended only for isometric objects, it also
does work for that case.